driver core: ipipe: defer dev_printk() from head domain
Just like printk(), dev_printk() cannot run from the head domain
and/or with hard IRQs disabled. In such a case, log the output
directly into the staging buffer we use for printk().

NOTE: when redirected to the buffer, the output does not include the
dev_printk() header text but is merely sent as-is to the log.