ipipe: Introduce infrastructure for userspace return notifier
A little bit inspired by the kernel's user return notifier, this
introduces an I-pipe hook before the kernel jumps back to a userspace
context from the root domain. The hook is design to allow a switch back
to the head domain, thus will not run through signal/preemption checks
when returning from the callback over head. It is guaranteed to fire on
return from interrupts and exceptions but may also fire on certain
syscall-return paths.

The first use case for the hook is resumption of ptraced tasks over
head if they were stopped in that domain.

This provides just the generic infrastructure, the invocation of
__ipipe_notify_user_intreturn as well as the definition of
TIP_USERINTRET are architecture-specific.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>