arm64: ipipe: unstall root stage at secondary CPU init
Turning on interrupts by flipping the DAIF bits is not enough when the
IRQ state is virtualized, we also have to inform the pipeline about
it, so that lockdep checks asserting IRQs on do not trigger.

Add an explicit call to local_irq_enable() prior to branching to
cpu_startup_entry(); this will do the necessary housekeeping bits such
as unstalling the root stage and telling the IRQ tracer about it.