arm64: ipipe: track IRQ state consistently upon BRK event
Two related issues to address:

1. Only after do_debug_exception() has told the irqsoff tracer about
the IRQ state on entry to the handler should we invoke
__ipipe_report_trap(), since the latter may have to switch the caller
to the root domain, unstalling it in the same move, which will
certainly affect the interrupt state.

2. the debug monitor handler should be called once the hardware and
virtual IRQ states are reconciled. To this end, enclose such call
inside the fault_entry/fault_exit section.