ipipe: Fix ipipe level irq end
The following commit in the vanilla kernel introduced
a check for the cached interrupt mask flag in mask_irq():

bf22ff45bed6 ("genirq: Avoid unnecessary low level irq function calls")

This means if the flag is not serviced correctly
the real bit in the hardware interrupt controller may not be
cleared or set.
The __ipipe_end_level_irq() function does not follow this rule.
It unmasks the bit in the hardware without setting the cached flags
accordingly. So after the first level interrupt is finished the
mask cache has a wrong state. If now the next interrupt fires,
the mask_irq() function will not really mask the interrupt in
the hardware which causes a interrupt storm after reenabeling
the hard irqs.
The fix now also updates the shadow flag correctly.

Signed-off-by: Gunter Grau <gunter.grau@philips.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Greg Gallagher <greg@embeddedgreg.com>