arm64: ipipe: Make erratum_1418040_thread_switch compatible with I-pipe
First of all, this erratum hook is called from __switch_to, thus
potentially also from the primary domain. Some of the functions it calls
check if preemption was disabled under Linux - which may not be the case
when invoked from primary domain. Rather than adding a costly check for
ipipe_root_p to this hot-path, simply turn the check off if I-pipe is
enabled.

As the hook can be called from primary context, we need to protect its
setup for new execs against those contexts via hard_preempt_disable.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Greg Gallagher <greg@embeddedgreg.com>